+++ src/cljs-lib/src/calva/repl/webview/core.cljs
@@ -142,14 +142,30 @@
   (.. vscode -window
       (onDidChangeActiveColorTheme
        (fn [e]
+         ;; Ensure syntax highlighting is not broken after the output view is moved into or out of the main
+         ;; VS Code window. See https://github.com/BetterThanTomorrow/calva/issues/2895.
+         ;; This is probably called more frequently than it needs to be, but if that's problematic we can try to
+         ;; figure out how to identify when this event means the output view moved into or out of the main window.
+         ;; This event fires for more cases than just that, such as when it's focused or unfocused.
          (set-code-theme! context {:color-theme-kind (.. e -kind)
                                    :webview-panel webview-panel})))))
+
+(defn create-view-state-change-listener
+  [{:keys [^js vscode/vscode] :as context}
+   {:keys [^js webview-panel]}]
+  (.. webview-panel
+      (onDidChangeViewState
+       (fn [^js _event]
+         (set-code-theme! context
+                          {:color-theme-kind (.. ^js vscode -window -activeColorTheme -kind)
+                           :webview-panel webview-panel})))))
 
 (defn add-subscriptions!
   [{vscode-context :vscode/context
     :as context}
    {:keys [webview-panel]}]
-  (let [subscriptions [(create-color-theme-change-listener context {:webview-panel webview-panel})]]
+  (let [subscriptions [(create-color-theme-change-listener context {:webview-panel webview-panel})
+                       (create-view-state-change-listener context {:webview-panel webview-panel})]]
     (run! (fn [subscription]
             (.. ^js vscode-context -subscriptions (push subscription)))
           subscriptions)))